list of contributors. untabified cu files-- fixed double literals that were causing nvcc to complain

+++ gpustats/cufiles/sampleFromMeasureMedium.cu
@@ -5,11 +5,9 @@
 
 
 __global__ void k_%(name)s(float* in_measure, /** Precomputed measure */
-			   float* in_random, /** Precomputed random number */
-			   int* out_component, /** Resultant choice */
-			   int iN, int iT) {
-  //const int iN = dims[0];
-  //const int iT = dims[1];
+						   float* in_random, /** Precomputed random number */
+						   int* out_component, /** Resultant choice */
+						   int iN, int iT) {
 
   const int sample_density_block = blockDim.x;
   const int sample_block = blockDim.y;
@@ -18,12 +16,8 @@ __global__ void k_%(name)s(float* in_measure, /** Precomputed measure */
   const int datumIndex = blockIdx.x * sample_block  + thidy;
   const int pdfIndex = datumIndex * iT;
   const int tid = thidy*sample_density_block + thidx;
- 
-  
-  //__shared__ REAL measure[sample_block][sample_density_block]; 
-  //__shared__ REAL sum[sample_block];
 
-  // Make block size flexible ... 
+  // Make block size flexible ...
   extern __shared__ float shared_data[];
   float* measure = shared_data; // sample_block by sample_density_block
   float* sum = measure + sample_block*sample_density_block;
@@ -34,7 +28,7 @@ __global__ void k_%(name)s(float* in_measure, /** Precomputed measure */
     sum[thidy] = 0;
 #if defined(LOGPDF)
     work[thidy] = -10000;
-#else 
+#else
     work[thidy] = 0;
 #endif
   }
@@ -44,15 +38,15 @@ __global__ void k_%(name)s(float* in_measure, /** Precomputed measure */
   for(int chunk = 0; chunk < iT; chunk += sample_density_block) {
     measure[thidy*sample_block + thidx] = in_measure[pdfIndex + chunk + thidx];
     __syncthreads();
-    
+
     if (thidx == 0) {
       for(int i=0; i<sample_density_block; i++) {
-	if(chunk + i < iT){
-	  float dcurrent = measure[thidy*sample_block + i];
-	  if (dcurrent > work[thidy]) {
-	    work[thidy] = dcurrent;
-	  }
-	}
+    if(chunk + i < iT){
+      float dcurrent = measure[thidy*sample_block + i];
+      if (dcurrent > work[thidy]) {
+        work[thidy] = dcurrent;
+      }
+    }
       }
     }
     __syncthreads();
@@ -64,26 +58,26 @@ __global__ void k_%(name)s(float* in_measure, /** Precomputed measure */
   for(int chunk = 0; chunk < iT; chunk += sample_density_block) {
 
     measure[thidy*sample_block + thidx] = in_measure[pdfIndex + chunk + thidx];
-    
+
     __syncthreads();
 
     if (thidx == 0) {
       for(int i=0; i<sample_density_block; i++) {
-	if (chunk + i < iT){
+    if (chunk + i < iT){
 #if defined(LOGPDF)
-	  //rescale and exp()
-	  sum[thidy] += exp(measure[thidy*sample_block + i] - work[thidy]); 
+      //rescale and exp()
+      sum[thidy] += expf(measure[thidy*sample_block + i] - work[thidy]);
 #else
-	  sum[thidy] += measure[thidy*sample_block + i];
+      sum[thidy] += measure[thidy*sample_block + i];
 #endif
-	  measure[thidy*sample_block + i] = sum[thidy];
-	}
+      measure[thidy*sample_block + i] = sum[thidy];
+    }
       }
     }
 
     if (chunk + thidx < iT)
       in_measure[pdfIndex + chunk + thidx] = measure[thidy*sample_block + thidx];
-    
+
     __syncthreads();
   }
 
@@ -102,20 +96,20 @@ __global__ void k_%(name)s(float* in_measure, /** Precomputed measure */
 
   // Find the right bin for the random number ...
   for(int chunk = 0; chunk < iT; chunk += sample_density_block) {
-    
+
     measure[thidy*sample_block + thidx] = in_measure[pdfIndex + chunk + thidx];
     __syncthreads();
 
     if (thidx == 0) {
 
       // storing the index in a float is better because it avoids
-      // bank conflicts ... 
+      // bank conflicts ...
       for(int i=0; i<sample_density_block; i++) {
-	if (chunk + i < iT){
-	  if (randomNumber[thidy] > measure[thidy*sample_block + i]){
-	    work[thidy] = i + chunk + 1;
-	  }
-	}
+    if (chunk + i < iT){
+      if (randomNumber[thidy] > measure[thidy*sample_block + i]){
+        work[thidy] = i + chunk + 1;
+      }
+    }
       }
       if ((int) work[thidy] >= iT) {work[thidy] = iT-1;}
     }
@@ -125,10 +119,6 @@ __global__ void k_%(name)s(float* in_measure, /** Precomputed measure */
   // this is now coalesced
   if (result_id < iN && tid < sample_block)
     out_component[result_id] = (int) work[tid];
-
-  //if (thidx == 0 && datumIndex < iN) 
-  //  out_component[datumIndex] = work;
-
 
 }
